nuevo módulo compartido para compartir componentes, objetos, modelos, etc
