+++ src/app/compartido/componentes/mangapreview/mangapreview.component.html
@@ -1,5 +1,5 @@
 <ion-card>
-  <img alt="{{ data.nombre }}" src="{{ data.rutaImg }}">
+  <img alt="{{ data.nombre }}" src="{{ data.enlaceImg }}">
   <ion-card-content>
     <ion-card-title>{{data.nombre}}</ion-card-title>
   </ion-card-content>
