+++ src/app/compartido/componentes/mangapreview/mangapreview.component.scss
