+++ src/app/compartido/componentes/mangapreview/mangapreview.component.ts
@@ -1,0 +1,17 @@
+import {Component, OnInit, Input} from '@angular/core';
+import {MangaPreviewModel} from '../../modelos/manga-preview.model';
+
+@Component({
+    selector: 'app-mangapreview',
+    templateUrl: './mangapreview.component.html',
+    styleUrls: ['./mangapreview.component.scss'],
+})
+export class MangapreviewComponent implements OnInit {
+    @Input() data: MangaPreviewModel;
+
+    constructor() {
+    }
+
+    ngOnInit() {
+    }
+}
