+++ src/app/compartido/modelos/manga-preview.model.spec.ts
@@ -1,0 +1,7 @@
+import { MangaPreviewModel } from './manga-preview.model';
+
+describe('MangaPreviewModel', () => {
+  it('should create an instance', () => {
+    expect(new MangaPreviewModel()).toBeTruthy();
+  });
+});
