+++ src/app/paginas/catalogo/catalogo.module.ts
@@ -7,7 +7,7 @@ import { IonicModule } from '@ionic/angular';
 import { CatalogoPageRoutingModule } from './catalogo-routing.module';
 
 import { CatalogoPage } from './catalogo.page';
-import {ComponentesCompartidosModule} from '../../componentes-compartidos/componentes-compartidos.module';
+import {CompartidoModule} from '../../compartido/compartido.module';
 
 @NgModule({
     imports: [
@@ -15,7 +15,7 @@ import {ComponentesCompartidosModule} from '../../componentes-compartidos/compon
         FormsModule,
         IonicModule,
         CatalogoPageRoutingModule,
-        ComponentesCompartidosModule
+        CompartidoModule
     ],
   declarations: [CatalogoPage]
 })
